+++ CHANGELOG.md
@@ -2,6 +2,10 @@
 
 ## Perf
 
+### 2025-10-13
+
+- Remove unnecesary mul in ecpairing [#4843](https://github.com/lambdaclass/ethrex/pull/4843)
+
 ### 2025-10-06
 
 - Improve block headers vec handling in syncer [#4771](https://github.com/lambdaclass/ethrex/pull/4771)